llm-product-recommendation-system/requirements.txt
FastAPI
SQLAlchemy
pydantic
uvicorn
transformers
torch
scikit-learn
python-dotenv
redis
pytest
httpx